requirements.txt
# Multi-agent crypto analytics stack

websockets>=12.0
aiohttp>=3.9.0
pandas>=2.1.0
numpy>=1.24.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Tests
pytest>=8.0.0
